refactor(angular2-textmask)!: replacing angular2-textmask usages
