+++ package.json
@@ -54,14 +54,12 @@
     "@bullhorn/dragula": "1.0.1",
     "ace-builds": "~1.4.12",
     "angular-imask": "6.2.2",
-    "angular2-text-mask": "9.0.0",
     "brace": "0.11.1",
     "date-fns": "^2.29.2",
     "imask": "^6.2.2",
     "novo-design-tokens": "^1.0.0-beta.6",
     "post-robot": "9.0.30",
     "rxjs": "^7.5.4",
-    "text-mask-addons": "^3.8.0",
     "timezone-support": "^2.0.2",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
