+++ projects/elements/components/date-time-picker/date-time-picker.module.ts
@@ -4,7 +4,6 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { IMaskDirectiveModule } from 'angular-imask';
 // Vendor
-import { TextMaskModule } from 'angular2-text-mask';
 import { NovoDateTimePickerInputElement } from './date-time-picker-input';
 import { NovoDateTimePickerElement } from './date-time-picker';
 import { NovoTimePickerModule } from 'novo-elements/components/time-picker';
@@ -12,7 +11,7 @@ import { NovoDatePickerModule } from 'novo-elements/components/date-picker';
 import { NovoOverlayModule } from 'novo-elements/common/overlay';
 
 @NgModule({
-  imports: [CommonModule, FormsModule, NovoDatePickerModule, NovoTimePickerModule, IMaskDirectiveModule, TextMaskModule, NovoOverlayModule],
+  imports: [CommonModule, FormsModule, NovoDatePickerModule, NovoTimePickerModule, IMaskDirectiveModule, NovoOverlayModule],
   declarations: [NovoDateTimePickerElement, NovoDateTimePickerInputElement],
   exports: [NovoDateTimePickerElement, NovoDateTimePickerInputElement],
 })
