+++ projects/examples/src/_shared/stackblitz/stackblitz-writer.ts
@@ -31,13 +31,11 @@ const dependencies = {
   rxjs: '6.5.5',
   '@bullhorn/dragula': '1.0.1',
   'ace-builds': '1.4.12',
-  'angular2-text-mask': '9.0.0',
   brace: '0.11.1',
   'classlist.js': '^1.1.20150312',
   'date-fns': '1.30.1',
   'post-robot': '9.0.30',
   'web-animations-js': '2.3.2',
-  'text-mask-addons': '^3.8.0',
   'zone.js': '^0.8.14',
   hammerjs: '^2.0.8',
   moment: '^2.18.1',
