+++ projects/examples/src/form-controls/date-picker/date-picker-input/date-picker-input-example.html
@@ -25,7 +25,7 @@
     <novo-radio name="weekStart" value="1">Mon</novo-radio>
   </novo-radio-group>
 
-  <novo-label>Week Start</novo-label>
+  <novo-label>Enabled</novo-label>
   <novo-radio-group #disabled appearance="vertical" [value]="false">
     <novo-radio name="disabled" [value]="false">Enabled</novo-radio>
     <novo-radio name="disabled" [value]="true">Disabled</novo-radio>
